Added product updates for action
