+++ apps/farcaster/components/common/Header.tsx
@@ -1,0 +1,8 @@
+'use client';
+
+import { useDatadogLogger } from '@root/hooks/useDatadogLogger';
+
+export function Header() {
+  useDatadogLogger({ service: 'farcaster-browser' });
+  return null;
+}
